Added app icon and tray icon
