+++ linux/gui/constants.py
@@ -1,0 +1,10 @@
+
+"""
+Monitorize GUI — Page index constants.
+"""
+
+PAGE_MAIN      = 0
+PAGE_WIFI      = 1
+PAGE_USB1      = 2
+PAGE_USB2      = 3
+PAGE_STREAMING = 4
