+++ linux/gui/pages/__init__.py
@@ -1,0 +1,9 @@
+
+"""
+Monitorize GUI — Page widgets.
+"""
+
+from gui.pages.main_menu import MainMenuPage
+from gui.pages.wifi_page import WifiPage
+from gui.pages.usb_pages import UsbStep1Page, UsbStep2Page
+from gui.pages.streaming_page import StreamingPage
