+++ linux/install.sh
@@ -1,0 +1,114 @@
+#!/usr/bin/env bash
+# ──────────────────────────────────────────────────────────────────────
+# Monitorize — Linux desktop installer
+#
+# Creates a .desktop entry so Monitorize appears in the application
+# menu on KDE, GNOME, Hyprland, and other freedesktop-compliant DEs.
+#
+# Usage:
+#   chmod +x install.sh
+#   ./install.sh          # install
+#   ./install.sh remove   # uninstall
+# ──────────────────────────────────────────────────────────────────────
+
+set -euo pipefail
+
+APP_NAME="Monitorize"
+APP_ID="monitorize"
+DESKTOP_FILE="${APP_ID}.desktop"
+
+# Resolve paths relative to this script (linux/ directory)
+SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
+ICON_SRC="${SCRIPT_DIR}/assets/monitorize-icon.png"
+ENTRY_POINT="${SCRIPT_DIR}/monitorize_gui.py"
+
+# XDG standard locations
+DESKTOP_DIR="${HOME}/.local/share/applications"
+ICON_DIR="${HOME}/.local/share/icons/hicolor/192x192/apps"
+ICON_DEST="${ICON_DIR}/${APP_ID}.png"
+
+# ── Uninstall ────────────────────────────────────────────────────────
+if [[ "${1:-}" == "remove" || "${1:-}" == "uninstall" ]]; then
+    echo "Removing ${APP_NAME} desktop entry…"
+    rm -f "${DESKTOP_DIR}/${DESKTOP_FILE}"
+    rm -f "${ICON_DEST}"
+    # Refresh desktop database if available
+    if command -v update-desktop-database &>/dev/null; then
+        update-desktop-database "${DESKTOP_DIR}" 2>/dev/null || true
+    fi
+    if command -v gtk-update-icon-cache &>/dev/null; then
+        gtk-update-icon-cache -f -t "${HOME}/.local/share/icons/hicolor" 2>/dev/null || true
+    fi
+    echo "✓ ${APP_NAME} has been removed from the application menu."
+    exit 0
+fi
+
+# ── Pre-flight checks ────────────────────────────────────────────────
+if [[ ! -f "${ICON_SRC}" ]]; then
+    echo "Error: Icon not found at ${ICON_SRC}" >&2
+    exit 1
+fi
+
+if [[ ! -f "${ENTRY_POINT}" ]]; then
+    echo "Error: Entry point not found at ${ENTRY_POINT}" >&2
+    exit 1
+fi
+
+# Check for python3
+if ! command -v python3 &>/dev/null; then
+    echo "Error: python3 is not installed." >&2
+    exit 1
+fi
+
+# ── Install icon ─────────────────────────────────────────────────────
+mkdir -p "${ICON_DIR}"
+cp "${ICON_SRC}" "${ICON_DEST}"
+echo "✓ Icon installed to ${ICON_DEST}"
+
+# ── Create .desktop file ─────────────────────────────────────────────
+mkdir -p "${DESKTOP_DIR}"
+
+cat > "${DESKTOP_DIR}/${DESKTOP_FILE}" <<EOF
+[Desktop Entry]
+Type=Application
+Name=${APP_NAME}
+Comment=Linux to Android Display Bridge — extend or mirror your desktop to a tablet
+Exec=python3 ${ENTRY_POINT}
+Icon=${APP_ID}
+Terminal=false
+Categories=Utility;System;
+Keywords=monitor;display;tablet;android;screen;extend;mirror;streaming;
+StartupNotify=true
+StartupWMClass=monitorize
+Path=${SCRIPT_DIR}
+EOF
+
+chmod +x "${DESKTOP_DIR}/${DESKTOP_FILE}"
+echo "✓ Desktop entry created at ${DESKTOP_DIR}/${DESKTOP_FILE}"
+
+# ── Refresh desktop database ─────────────────────────────────────────
+if command -v update-desktop-database &>/dev/null; then
+    update-desktop-database "${DESKTOP_DIR}" 2>/dev/null || true
+    echo "✓ Desktop database updated"
+fi
+
+# Refresh icon cache so DEs pick up the new icon immediately
+if command -v gtk-update-icon-cache &>/dev/null; then
+    gtk-update-icon-cache -f -t "${HOME}/.local/share/icons/hicolor" 2>/dev/null || true
+    echo "✓ Icon cache updated"
+fi
+
+# Pre-compile Python files to .pyc to ensure maximum startup speed
+python3 -m compileall "${SCRIPT_DIR}" &>/dev/null || true
+echo "✓ Python bytecode pre-compiled"
+
+echo ""
+echo "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
+echo "  ${APP_NAME} has been installed!"
+echo "  It should now appear in your application menu."
+echo ""
+echo "  If it doesn't appear immediately, try logging out and"
+echo "  back in, or run:  kbuildsycoca6  (KDE)"
+echo ""
+echo "  To uninstall:  ./install.sh remove"
+echo "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
