+++ linux/monitorize_gui.py
@@ -2,1879 +2,17 @@
 """
 Monitorize GUI — PyQt6 control panel
 Run from the linux/ directory:  python3 monitorize_gui.py
+
+This file is a thin launcher that delegates to the gui/ package.
 """
 
 import sys
 import os
-from PyQt6.QtWidgets import (
-    QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout,
-    QLabel, QPushButton, QStackedWidget, QFrame, QPlainTextEdit,
-    QComboBox, QCheckBox, QSystemTrayIcon, QMenu,
-    QDialog, QMessageBox, QLineEdit, QScrollArea,
-)
-from PyQt6.QtCore import Qt, QProcess, QProcessEnvironment, QTimer, QSize, QSettings
-from PyQt6.QtGui import QColor, QPalette, QFont, QTextCursor, QIcon, QPixmap, QPainter
 
 
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
-
-
-DARK_QSS = """
-/* ── Base ─────────────────────────────────────────────────────────── */
-QMainWindow, QWidget {
-    background-color: #0c0d14;
-    color: #d4d6f0;
-    font-family: 'Inter', 'SF Pro Display', 'Segoe UI', sans-serif;
-    font-size: 14px;
-}
-
-/* ── Generic Button ───────────────────────────────────────────────── */
-QPushButton {
-    background-color: #16182a;
-    color: #b8bad8;
-    border: 1px solid #252845;
-    border-radius: 10px;
-    padding: 10px 24px;
-    font-size: 14px;
-    font-weight: 600;
-}
-QPushButton:hover {
-    background-color: #1e2040;
-    border-color: #4a4faa;
-    color: #e8e9ff;
-}
-QPushButton:pressed { background-color: #121428; }
-QPushButton:disabled {
-    background-color: #101220;
-    color: #3e3f5a;
-    border-color: #1a1c30;
-}
-
-/* ── Mode Buttons (USB / Wi-Fi) ───────────────────────────────────── */
-QPushButton#modeBtn {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #141630, stop:1 #1c1e3a);
-    border: 1px solid #2a2d55;
-    border-radius: 16px;
-    font-size: 18px;
-    font-weight: 700;
-    color: #c0c2ee;
-    min-width: 200px;
-    min-height: 110px;
-}
-QPushButton#modeBtn:hover {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #1c1e42, stop:1 #24264e);
-    border-color: #5458b8;
-    color: #ffffff;
-}
-
-/* ── Primary Action Button ────────────────────────────────────────── */
-QPushButton#primaryBtn {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #3538b0, stop:1 #4c4fd0);
-    border: none;
-    border-radius: 12px;
-    font-size: 15px;
-    font-weight: 700;
-    color: #ffffff;
-    min-height: 50px;
-    min-width: 240px;
-}
-QPushButton#primaryBtn:hover {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #4042c8, stop:1 #5c5ee8);
-}
-QPushButton#primaryBtn:pressed { background-color: #2a2c98; }
-
-/* ── Stop Button ──────────────────────────────────────────────────── */
-QPushButton#stopBtn {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #7a1520, stop:1 #a82028);
-    border: none;
-    border-radius: 12px;
-    font-size: 15px;
-    font-weight: 700;
-    color: #ffffff;
-    min-height: 50px;
-    min-width: 240px;
-}
-QPushButton#stopBtn:hover {
-    background: qlineargradient(x1:0,y1:0,x2:1,y2:1,
-                stop:0 #941a24, stop:1 #c42830);
-}
-QPushButton#stopBtn:pressed { background-color: #5a1010; }
-
-/* ── Back Button ──────────────────────────────────────────────────── */
-QPushButton#backBtn {
-    background-color: transparent;
-    border: 1px solid #252845;
-    border-radius: 10px;
-    font-size: 13px;
-    color: #6a6c90;
-    padding: 8px 18px;
-    min-width: 90px;
-}
-QPushButton#backBtn:hover { border-color: #4a4faa; color: #b0b2d8; }
-
-/* ── Labels ───────────────────────────────────────────────────────── */
-QLabel#titleLabel {
-    font-size: 30px;
-    font-weight: 800;
-    color: #e0e2ff;
-    letter-spacing: 2px;
-}
-QLabel#subLabel    { font-size: 14px; color: #6a6c96; font-weight: 400; }
-QLabel#stepLabel   { font-size: 12px; color: #5a5c82; font-weight: 500; letter-spacing: 1px; }
-QLabel#instruction { font-size: 15px; color: #b0b2d0; }
-QLabel#wip         { font-size: 17px; color: #7878aa; font-style: italic; }
-QLabel#streaming   { font-size: 20px; font-weight: 700; color: #4cd68d; }
-QLabel#statusLbl   { font-size: 12px; color: #5a5c82; }
-
-/* ── Log Box ──────────────────────────────────────────────────────── */
-QPlainTextEdit#logBox {
-    background-color: #080910;
-    color: #7cc87c;
-    border: 1px solid #1a1c30;
-    border-radius: 8px;
-    font-family: 'JetBrains Mono', 'Fira Code', 'Consolas', monospace;
-    font-size: 11px;
-    padding: 8px;
-    selection-background-color: #2a2d55;
-}
-
-/* ── Separator ────────────────────────────────────────────────────── */
-QFrame#sep {
-    background-color: #1a1c30;
-    max-height: 1px;
-}
-
-/* ── Combo Box ────────────────────────────────────────────────────── */
-QComboBox {
-    background-color: #12142a;
-    color: #c0c2ee;
-    border: 1px solid #2a2d55;
-    border-radius: 8px;
-    padding: 8px 14px;
-    font-size: 13px;
-    font-weight: 600;
-    min-width: 140px;
-}
-QComboBox:hover { border-color: #5458b8; }
-QComboBox::drop-down { border: none; width: 24px; }
-QComboBox QAbstractItemView {
-    background-color: #12142a;
-    color: #c0c2ee;
-    selection-background-color: #3538b0;
-    selection-color: #ffffff;
-    border: 1px solid #2a2d55;
-    border-radius: 6px;
-    padding: 4px;
-}
-
-/* ── Custom Input ─────────────────────────────────────────────────── */
-QLineEdit#customInput {
-    background-color: #12142a;
-    color: #c0c2ee;
-    border: 1px solid #2a2d55;
-    border-radius: 8px;
-    padding: 8px 10px;
-    font-size: 13px;
-    font-weight: 600;
-    min-width: 80px;
-    max-width: 90px;
-}
-QLineEdit#customInput:focus { border-color: #5458b8; }
-QLineEdit#customInput[invalid="true"] {
-    border-color: #a82028;
-    color: #ff6060;
-}
-
-QLabel#xSep {
-    font-size: 18px;
-    font-weight: 700;
-    color: #6a6c96;
-    padding: 0 4px;
-}
-QLabel#customHint {
-    font-size: 11px;
-    color: #4a4c70;
-    font-style: italic;
-}
-
-/* ── Warning Label ────────────────────────────────────────────────── */
-QLabel#warningLabel {
-    font-size: 12px;
-    font-weight: 600;
-    color: #e8a840;
-    padding: 8px 14px;
-    background-color: rgba(232, 168, 64, 0.06);
-    border: 1px solid rgba(232, 168, 64, 0.18);
-    border-radius: 8px;
-}
-
-/* ── DE Badge ─────────────────────────────────────────────────────── */
-#deBadge {
-    background-color: rgba(76, 79, 208, 0.08);
-    border: 1px solid rgba(76, 79, 208, 0.16);
-    border-radius: 14px;
-}
-#deBadge QLabel {
-    font-size: 12px;
-    font-weight: 600;
-    color: #6a6cbb;
-    background: transparent;
-    border: none;
-}
-
-/* ── Portal Hint ──────────────────────────────────────────────────── */
-QLabel#portalHint {
-    font-size: 14px;
-    font-weight: 500;
-    color: #8a8cc0;
-}
-
-/* ── Tray Checkbox ────────────────────────────────────────────────── */
-QCheckBox#trayCheck {
-    font-size: 12px;
-    color: #5a5c82;
-    spacing: 8px;
-}
-QCheckBox#trayCheck:hover { color: #9a9cc0; }
-QCheckBox#trayCheck::indicator {
-    width: 14px;
-    height: 14px;
-    border: 1px solid #2a2d55;
-    border-radius: 4px;
-    background-color: #12142a;
-}
-QCheckBox#trayCheck::indicator:checked {
-    background-color: #4c4fd0;
-    border-color: #4c4fd0;
-}
-
-/* ── Scrollbar ────────────────────────────────────────────────────── */
-QScrollBar:vertical {
-    background: transparent;
-    width: 6px;
-    margin: 0;
-}
-QScrollBar::handle:vertical {
-    background: #2a2d55;
-    border-radius: 3px;
-    min-height: 30px;
-}
-QScrollBar::handle:vertical:hover { background: #3a3d6a; }
-QScrollBar::add-line:vertical, QScrollBar::sub-line:vertical { height: 0; }
-QScrollBar::add-page:vertical, QScrollBar::sub-page:vertical { background: transparent; }
-
-/* ── Message Boxes / Dialogs ──────────────────────────────────────── */
-QDialog {
-    background-color: #0c0d14;
-}
-QMessageBox {
-    background-color: #0c0d14;
-}
-QMessageBox QLabel {
-    color: #d4d6f0;
-}
-"""
-
-
-
-
-
-class NonScrollComboBox(QComboBox):
-    """A QComboBox that ignores mouse wheel events to prevent accidental setting changes while scrolling the panel."""
-    def wheelEvent(self, event):
-        event.ignore()
-
-def hr() -> QFrame:
-    line = QFrame()
-    line.setObjectName("sep")
-    line.setFrameShape(QFrame.Shape.HLine)
-    return line
-
-def vspace(n: int) -> int:
-    return n  
-
-def _make_tray_icon() -> QIcon:
-    """Generate a simple coloured square icon for the system tray."""
-    px = QPixmap(64, 64)
-    px.fill(Qt.GlobalColor.transparent)
-    p = QPainter(px)
-    p.setRenderHint(QPainter.RenderHint.Antialiasing)
-    p.setBrush(QColor("#4c4fd0"))
-    p.setPen(Qt.PenStyle.NoPen)
-    p.drawEllipse(4, 4, 56, 56)
-    p.setBrush(QColor("#ffffff"))
-    p.drawEllipse(20, 20, 24, 24)
-    p.end()
-    return QIcon(px)
-
-
-
-
-
-
-def detect_desktop_environment() -> str:
-    """
-    Return "kde", "gnome", "hyprland", "sway", or "" (unknown) based on
-    environment variables.  Checks XDG_CURRENT_DESKTOP, DESKTOP_SESSION,
-    and the Hyprland/Sway-specific vars; case-insensitive.
-    """
-    xdg   = os.environ.get("XDG_CURRENT_DESKTOP", "").lower()
-    dsess = os.environ.get("DESKTOP_SESSION",      "").lower()
-    
-    hypr  = os.environ.get("HYPRLAND_INSTANCE_SIGNATURE", "")
-    sway  = os.environ.get("SWAYSOCK", "")
-    combined = xdg + " " + dsess
-
-    if hypr or "hyprland" in combined:
-        return "hyprland"
-    if sway or "sway" in combined:
-        return "sway"
-    if "kde" in combined:
-        return "kde"
-    if "gnome" in combined:
-        return "gnome"
-    return ""
-
-
-
-
-
-
-class MainMenuPage(QWidget):
-    def __init__(self, on_usb, on_wifi, parent=None):
-        super().__init__(parent)
-        root = QVBoxLayout(self)
-        root.setContentsMargins(50, 50, 50, 40)
-
-        title = QLabel("Monitorize")
-        title.setObjectName("titleLabel")
-        title.setAlignment(Qt.AlignmentFlag.AlignCenter)
-
-        sub = QLabel("Linux → Android Display Bridge")
-        sub.setObjectName("subLabel")
-        sub.setAlignment(Qt.AlignmentFlag.AlignCenter)
-
-        
-        self._de_badge = QFrame()
-        self._de_badge.setObjectName("deBadge")
-        badge_layout = QHBoxLayout(self._de_badge)
-        badge_layout.setContentsMargins(14, 4, 14, 4)
-        badge_layout.setSpacing(6)
-        
-        self._de_badge_icon = QLabel()
-        self._de_badge_text = QLabel("")
-        badge_layout.addStretch()
-        badge_layout.addWidget(self._de_badge_icon)
-        badge_layout.addWidget(self._de_badge_text)
-        badge_layout.addStretch()
-
-        root.addWidget(title)
-        root.addSpacing(6)
-        root.addWidget(sub)
-        root.addSpacing(10)
-        root.addWidget(self._de_badge)
-        root.addSpacing(24)
-        root.addWidget(hr())
-        root.addSpacing(44)
-
-        row = QHBoxLayout()
-        row.setSpacing(28)
-
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        usb_svg = os.path.join(script_dir, "assets", "svg", "usb-logo.svg")
-        wifi_svg = os.path.join(script_dir, "assets", "svg", "wifi-logo.svg")
-
-        usb_btn = QPushButton("  USB Mode")
-        usb_btn.setObjectName("modeBtn")
-        if os.path.exists(usb_svg):
-            usb_btn.setIcon(QIcon(usb_svg))
-            usb_btn.setIconSize(QSize(20, 20))
-        usb_btn.clicked.connect(on_usb)
-
-        wifi_btn = QPushButton("  Wi-Fi Mode")
-        wifi_btn.setObjectName("modeBtn")
-        if os.path.exists(wifi_svg):
-            wifi_btn.setIcon(QIcon(wifi_svg))
-            wifi_btn.setIconSize(QSize(20, 20))
-        wifi_btn.clicked.connect(on_wifi)
-
-        row.addStretch()
-        row.addWidget(usb_btn)
-        row.addWidget(wifi_btn)
-        row.addStretch()
-        root.addLayout(row)
-        root.addStretch()
-
-        footer = QLabel("Select a connection mode to begin")
-        footer.setObjectName("statusLbl")
-        footer.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        root.addWidget(footer)
-        root.addSpacing(14)
-
-        
-        tray_row = QHBoxLayout()
-        self.tray_checkbox = QCheckBox("Minimize to tray on close")
-        self.tray_checkbox.setObjectName("trayCheck")
-        self.tray_checkbox.setChecked(False)
-        tray_row.addStretch()
-        tray_row.addWidget(self.tray_checkbox)
-        tray_row.addStretch()
-        root.addLayout(tray_row)
-        root.addSpacing(6)
-
-    def update_de_badge(self, de: str):
-        """Show the detected/selected desktop environment in the badge."""
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        _icons = {
-            "kde":      os.path.join(script_dir, "assets", "svg", "kde-logo.svg"),
-            "gnome":    os.path.join(script_dir, "assets", "svg", "gnome-logo.svg"),
-            "hyprland": os.path.join(script_dir, "assets", "svg", "hyprland-logo.svg"),
-        }
-        _labels = {
-            "kde":      "KDE Plasma",
-            "gnome":    "GNOME",
-            "hyprland": "Hyprland",
-            "sway":     "Sway",
-        }
-        icon_path = _icons.get(de)
-        if icon_path and os.path.exists(icon_path):
-            self._de_badge_icon.setPixmap(QIcon(icon_path).pixmap(16, 16))
-            self._de_badge_icon.setVisible(True)
-        else:
-            self._de_badge_icon.setVisible(False)
-
-        label = _labels.get(de, de.upper() if de else "Unknown")
-        self._de_badge_text.setText(f"Desktop: {label}")
-
-
-def get_local_ip():
-    import socket
-    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    try:
-        s.connect(('10.255.255.255', 1))
-        IP = s.getsockname()[0]
-    except Exception:
-        IP = '127.0.0.1'
-    finally:
-        s.close()
-    return IP
-
-class WifiPage(QWidget):
-    RESOLUTIONS = [
-        "1280x720", "1280x800", "1920x1080", "1920x1200",
-        "2560x1440", "2560x1600", "3840x2160", "Custom…"
-    ]
-    FPS_OPTIONS = ["30", "60", "90", "120", "Custom…"]
-
-    def __init__(self, on_back, on_start, detected_de="kde", parent=None):
-        super().__init__(parent)
-        self._on_start_cb = on_start
-        root = QVBoxLayout(self)
-        root.setContentsMargins(60, 40, 60, 40)
-
-        step = QLabel("Wi-Fi Mode")
-        step.setObjectName("stepLabel")
-        root.addWidget(step)
-        root.addSpacing(12)
-        root.addWidget(hr())
-        root.addSpacing(24)
-
-        ip = get_local_ip()
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        wifi_svg = os.path.join(script_dir, "assets", "svg", "wifi-logo.svg")
-
-        wifi_header = QWidget()
-        wifi_header_layout = QHBoxLayout(wifi_header)
-        wifi_header_layout.setContentsMargins(0, 0, 0, 0)
-        wifi_header_layout.setSpacing(8)
-
-        wifi_icon_lbl = QLabel()
-        if os.path.exists(wifi_svg):
-            wifi_icon_lbl.setPixmap(QIcon(wifi_svg).pixmap(24, 24))
-
-        wifi_text_lbl = QLabel(f"Your Local IP Address is: {ip}")
-        wifi_text_lbl.setObjectName("instruction")
-        wifi_text_lbl.setStyleSheet("font-size: 16px; font-weight: bold;")
-
-        wifi_header_layout.addStretch()
-        wifi_header_layout.addWidget(wifi_icon_lbl)
-        wifi_header_layout.addWidget(wifi_text_lbl)
-        wifi_header_layout.addStretch()
-        root.addWidget(wifi_header)
-
-        msg = QLabel("Enter this IP in the Monitorize Android app and tap Receive.")
-        msg.setObjectName("instruction")
-        msg.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        msg.setWordWrap(True)
-        msg.setStyleSheet("font-size: 14px; color: #8a8cc0;")
-        root.addWidget(msg)
-        root.addSpacing(20)
-
-        
-        res_row = QHBoxLayout()
-        res_row.setSpacing(12)
-        res_label = QLabel("Resolution:")
-        res_label.setObjectName("instruction")
-        self._res_combo = NonScrollComboBox()
-        self._res_combo.addItems(self.RESOLUTIONS)
-        self._res_combo.setCurrentText("2560x1600")
-        self._res_combo.currentTextChanged.connect(self._on_res_changed)
-        res_row.addStretch()
-        res_row.addWidget(res_label)
-        res_row.addWidget(self._res_combo)
-        res_row.addStretch()
-        root.addLayout(res_row)
-
-        self._custom_res_widget = QWidget()
-        custom_res_inner = QHBoxLayout(self._custom_res_widget)
-        custom_res_inner.setContentsMargins(0, 4, 0, 0)
-        custom_res_inner.setSpacing(6)
-        self._custom_w = QLineEdit()
-        self._custom_w.setObjectName("customInput")
-        self._custom_w.setPlaceholderText("Width")
-        self._custom_w.setMaxLength(4)
-        x_sep = QLabel("×")
-        x_sep.setObjectName("xSep")
-        self._custom_h = QLineEdit()
-        self._custom_h.setObjectName("customInput")
-        self._custom_h.setPlaceholderText("Height")
-        self._custom_h.setMaxLength(4)
-        res_hint = QLabel("(500 – 4000 each)")
-        res_hint.setObjectName("customHint")
-        custom_res_inner.addStretch()
-        custom_res_inner.addWidget(self._custom_w)
-        custom_res_inner.addWidget(x_sep)
-        custom_res_inner.addWidget(self._custom_h)
-        custom_res_inner.addWidget(res_hint)
-        custom_res_inner.addStretch()
-        self._custom_res_widget.setVisible(False)
-        root.addWidget(self._custom_res_widget)
-        root.addSpacing(10)
-
-        
-        fps_row = QHBoxLayout()
-        fps_row.setSpacing(12)
-        fps_label = QLabel("FPS:")
-        fps_label.setObjectName("instruction")
-        self._fps_combo = NonScrollComboBox()
-        self._fps_combo.addItems(self.FPS_OPTIONS)
-        self._fps_combo.setCurrentText("60")
-        self._fps_combo.currentTextChanged.connect(self._on_fps_changed)
-        fps_row.addStretch()
-        fps_row.addWidget(fps_label)
-        fps_row.addWidget(self._fps_combo)
-        fps_row.addStretch()
-        root.addLayout(fps_row)
-
-        self._custom_fps_widget = QWidget()
-        custom_fps_inner = QHBoxLayout(self._custom_fps_widget)
-        custom_fps_inner.setContentsMargins(0, 4, 0, 0)
-        custom_fps_inner.setSpacing(6)
-        self._custom_fps_edit = QLineEdit()
-        self._custom_fps_edit.setObjectName("customInput")
-        self._custom_fps_edit.setPlaceholderText("FPS")
-        self._custom_fps_edit.setMaxLength(3)
-        fps_hint = QLabel("(24 – 240)")
-        fps_hint.setObjectName("customHint")
-        custom_fps_inner.addStretch()
-        custom_fps_inner.addWidget(self._custom_fps_edit)
-        custom_fps_inner.addWidget(fps_hint)
-        custom_fps_inner.addStretch()
-        self._custom_fps_widget.setVisible(False)
-        root.addWidget(self._custom_fps_widget)
-        root.addSpacing(10)
-
-        
-        bitrate_row = QHBoxLayout()
-        bitrate_row.setSpacing(12)
-        bitrate_label = QLabel("Video Bitrate (kbps):")
-        bitrate_label.setObjectName("instruction")
-        self._bitrate_edit = QLineEdit()
-        self._bitrate_edit.setObjectName("customInput")
-        self._bitrate_edit.setText("8000")
-        self._bitrate_edit.setMaxLength(5)
-        bitrate_row.addStretch()
-        bitrate_row.addWidget(bitrate_label)
-        bitrate_row.addWidget(self._bitrate_edit)
-        bitrate_row.addStretch()
-        root.addLayout(bitrate_row)
-        root.addSpacing(16)
-
-        self._show_display_type = (detected_de in ("gnome", "kde", "hyprland"))
-        if self._show_display_type:
-            type_row = QHBoxLayout()
-            type_row.setSpacing(20)
-            
-            type_lbl = QLabel("Display Type:")
-            type_lbl.setObjectName("instruction")
-            self._display_type_combo = NonScrollComboBox()
-            self._display_type_combo.addItems(["Extend Right", "Mirror"])
-            self._display_type_combo.setCurrentText("Extend Right")
-            
-            type_row.addStretch()
-            type_row.addWidget(type_lbl)
-            type_row.addWidget(self._display_type_combo)
-            type_row.addStretch()
-            
-            root.addLayout(type_row)
-            root.addSpacing(16)
-
-        warning = QLabel(
-            "WARNING: The Resolution and FPS set here MUST EXACTLY "
-            "MATCH the settings in the Android tablet app, or the stream "
-            "will corrupt!"
-        )
-        warning.setObjectName("warningLabel")
-        warning.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        warning.setWordWrap(True)
-        root.addWidget(warning)
-        root.addSpacing(20)
-
-        root.addWidget(hr())
-        root.addSpacing(16)
-
-        step2_lbl = QLabel(
-            'Then click Start Streaming below. When the screen-sharing '
-            'popup appears, select "TabletDisplay" and click Share.'
-        )
-        step2_lbl.setObjectName("portalHint")
-        step2_lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        step2_lbl.setWordWrap(True)
-        root.addWidget(step2_lbl)
-        root.addSpacing(20)
-
-        root.addStretch()
-
-        back = QPushButton("← Back")
-        back.setObjectName("backBtn")
-        back.clicked.connect(on_back)
-
-        self._start_btn = QPushButton("▶  Start Streaming")
-        self._start_btn.setObjectName("primaryBtn")
-        self._start_btn.clicked.connect(self._validate_and_start)
-
-        row = QHBoxLayout()
-        row.setSpacing(20)
-        row.addStretch()
-        row.addWidget(back)
-        row.addWidget(self._start_btn)
-        row.addStretch()
-        root.addLayout(row)
-        root.addSpacing(20)
-
-    def _on_res_changed(self, text: str):
-        self._custom_res_widget.setVisible(text == "Custom…")
-
-    def _on_fps_changed(self, text: str):
-        self._custom_fps_widget.setVisible(text == "Custom…")
-
-    def _validate_and_start(self):
-        if self._res_combo.currentText() == "Custom…":
-            try:
-                w = int(self._custom_w.text())
-                h = int(self._custom_h.text())
-            except ValueError: return
-        if self._fps_combo.currentText() == "Custom…":
-            try: fps = int(self._custom_fps_edit.text())
-            except ValueError: return
-        try: bitrate = int(self._bitrate_edit.text())
-        except ValueError: return
-
-        self._on_start_cb()
-
-    def selected_resolution(self) -> tuple[int, int]:
-        if self._res_combo.currentText() == "Custom…":
-            return int(self._custom_w.text()), int(self._custom_h.text())
-        text = self._res_combo.currentText()
-        w, h = text.split("x")
-        return int(w), int(h)
-
-    def selected_fps(self) -> int:
-        if self._fps_combo.currentText() == "Custom…":
-            return int(self._custom_fps_edit.text())
-        return int(self._fps_combo.currentText())
-
-    def selected_bitrate(self) -> int:
-        return int(self._bitrate_edit.text())
-
-    def gnome_scale(self) -> str:
-        if hasattr(self, "_gnome_scale_combo"):
-            return self._gnome_scale_combo.currentText()
-        return "1.0"
-
-    def gnome_type(self) -> str:
-        if hasattr(self, "_display_type_combo"):
-            return self._display_type_combo.currentText()
-        return "Extend Right"
-
-class UsbStep1Page(QWidget):
-    """Step 1 — connect tablet and run ADB."""
-
-    def __init__(self, on_back, on_connected, parent=None):
-        super().__init__(parent)
-        root = QVBoxLayout(self)
-        root.setContentsMargins(60, 60, 60, 60)
-
-        step = QLabel("USB Mode  ·  Step 1 of 2")
-        step.setObjectName("stepLabel")
-        root.addWidget(step)
-        root.addSpacing(12)
-        root.addWidget(hr())
-        root.addStretch()
-
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        usb_svg = os.path.join(script_dir, "assets", "svg", "usb-logo.svg")
-
-        icon = QLabel()
-        icon.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        if os.path.exists(usb_svg):
-            icon.setPixmap(QIcon(usb_svg).pixmap(96, 96))
-        root.addWidget(icon)
-        root.addSpacing(24)
-
-        msg = QLabel(
-            "Please connect your tablet via USB\n"
-            "with ADB debugging enabled."
-        )
-        msg.setObjectName("instruction")
-        msg.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        msg.setWordWrap(True)
-        root.addWidget(msg)
-        root.addStretch()
-
-        self._status = QLabel("")
-        self._status.setObjectName("statusLbl")
-        self._status.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        root.addWidget(self._status)
-        root.addSpacing(16)
-
-        self._next_btn = QPushButton("I have connected it")
-        self._next_btn.setObjectName("primaryBtn")
-        self._next_btn.clicked.connect(on_connected)
-
-        back = QPushButton("← Back")
-        back.setObjectName("backBtn")
-        back.clicked.connect(on_back)
-
-        row = QHBoxLayout()
-        row.setSpacing(20)
-        row.addStretch()
-        row.addWidget(back)
-        row.addWidget(self._next_btn)
-        row.addStretch()
-        root.addLayout(row)
-        root.addSpacing(24)
-
-    def set_status(self, text: str):
-        self._status.setText(text)
-
-    def set_busy(self, busy: bool):
-        self._next_btn.setEnabled(not busy)
-
-
-class UsbStep2Page(QWidget):
-    """Step 2 — configure resolution/FPS, open tablet app, start streaming."""
-
-    RESOLUTIONS = [
-        "1280x720", "1280x800", "1920x1080", "1920x1200",
-        "2560x1440", "2560x1600", "3840x2160", "Custom…",
-    ]
-    FPS_OPTIONS = ["30", "60", "90", "120", "Custom…"]
-
-    def __init__(self, on_back, on_start, detected_de="kde", parent=None):
-        super().__init__(parent)
-        self._on_start_cb = on_start
-        root = QVBoxLayout(self)
-        root.setContentsMargins(60, 40, 60, 40)
-
-        step = QLabel("USB Mode  ·  Step 2 of 2")
-        step.setObjectName("stepLabel")
-        root.addWidget(step)
-        root.addSpacing(12)
-        root.addWidget(hr())
-        root.addSpacing(24)
-
-        
-        msg = QLabel("Please open the Monitorize app on your tablet.")
-        msg.setObjectName("instruction")
-        msg.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        msg.setWordWrap(True)
-        root.addWidget(msg)
-        root.addSpacing(20)
-
-        
-        res_row = QHBoxLayout()
-        res_row.setSpacing(12)
-        res_label = QLabel("Resolution:")
-        res_label.setObjectName("instruction")
-        self._res_combo = NonScrollComboBox()
-        self._res_combo.addItems(self.RESOLUTIONS)
-        self._res_combo.setCurrentText("2560x1600")
-        self._res_combo.currentTextChanged.connect(self._on_res_changed)
-        res_row.addStretch()
-        res_row.addWidget(res_label)
-        res_row.addWidget(self._res_combo)
-        res_row.addStretch()
-        root.addLayout(res_row)
-
-        
-        self._custom_res_widget = QWidget()
-        custom_res_inner = QHBoxLayout(self._custom_res_widget)
-        custom_res_inner.setContentsMargins(0, 4, 0, 0)
-        custom_res_inner.setSpacing(6)
-        self._custom_w = QLineEdit()
-        self._custom_w.setObjectName("customInput")
-        self._custom_w.setPlaceholderText("Width")
-        self._custom_w.setMaxLength(4)
-        x_sep = QLabel("×")
-        x_sep.setObjectName("xSep")
-        self._custom_h = QLineEdit()
-        self._custom_h.setObjectName("customInput")
-        self._custom_h.setPlaceholderText("Height")
-        self._custom_h.setMaxLength(4)
-        res_hint = QLabel("(500 – 4000 each)")
-        res_hint.setObjectName("customHint")
-        custom_res_inner.addStretch()
-        custom_res_inner.addWidget(self._custom_w)
-        custom_res_inner.addWidget(x_sep)
-        custom_res_inner.addWidget(self._custom_h)
-        custom_res_inner.addWidget(res_hint)
-        custom_res_inner.addStretch()
-        self._custom_res_widget.setVisible(False)
-        root.addWidget(self._custom_res_widget)
-        root.addSpacing(10)
-
-        
-        fps_row = QHBoxLayout()
-        fps_row.setSpacing(12)
-        fps_label = QLabel("FPS:")
-        fps_label.setObjectName("instruction")
-        self._fps_combo = NonScrollComboBox()
-        self._fps_combo.addItems(self.FPS_OPTIONS)
-        self._fps_combo.setCurrentText("60")
-        self._fps_combo.currentTextChanged.connect(self._on_fps_changed)
-        fps_row.addStretch()
-        fps_row.addWidget(fps_label)
-        fps_row.addWidget(self._fps_combo)
-        fps_row.addStretch()
-        root.addLayout(fps_row)
-
-        
-        self._custom_fps_widget = QWidget()
-        custom_fps_inner = QHBoxLayout(self._custom_fps_widget)
-        custom_fps_inner.setContentsMargins(0, 4, 0, 0)
-        custom_fps_inner.setSpacing(6)
-        self._custom_fps_edit = QLineEdit()
-        self._custom_fps_edit.setObjectName("customInput")
-        self._custom_fps_edit.setPlaceholderText("FPS")
-        self._custom_fps_edit.setMaxLength(3)
-        fps_hint = QLabel("(24 – 240)")
-        fps_hint.setObjectName("customHint")
-        custom_fps_inner.addStretch()
-        custom_fps_inner.addWidget(self._custom_fps_edit)
-        custom_fps_inner.addWidget(fps_hint)
-        custom_fps_inner.addStretch()
-        self._custom_fps_widget.setVisible(False)
-        root.addWidget(self._custom_fps_widget)
-        root.addSpacing(10)
-
-        
-        bitrate_row = QHBoxLayout()
-        bitrate_row.setSpacing(12)
-        bitrate_label = QLabel("Video Bitrate (kbps):")
-        bitrate_label.setObjectName("instruction")
-        self._bitrate_edit = QLineEdit()
-        self._bitrate_edit.setObjectName("customInput")
-        self._bitrate_edit.setText("8000")
-        self._bitrate_edit.setMaxLength(5)
-        bitrate_row.addStretch()
-        bitrate_row.addWidget(bitrate_label)
-        bitrate_row.addWidget(self._bitrate_edit)
-        bitrate_row.addStretch()
-        root.addLayout(bitrate_row)
-        root.addSpacing(16)
-
-        self._show_display_type = (detected_de in ("gnome", "kde", "hyprland"))
-        if self._show_display_type:
-            type_row = QHBoxLayout()
-            type_row.setSpacing(20)
-            
-            type_lbl = QLabel("Display Type:")
-            type_lbl.setObjectName("instruction")
-            self._display_type_combo = NonScrollComboBox()
-            self._display_type_combo.addItems(["Extend Right", "Mirror"])
-            self._display_type_combo.setCurrentText("Extend Right")
-            
-            type_row.addStretch()
-            type_row.addWidget(type_lbl)
-            type_row.addWidget(self._display_type_combo)
-            type_row.addStretch()
-            
-            root.addLayout(type_row)
-            root.addSpacing(16)
-
-        
-        warning = QLabel(
-            "WARNING: The Resolution and FPS set here MUST EXACTLY "
-            "MATCH the settings in the Android tablet app, or the stream "
-            "will corrupt!"
-        )
-        warning.setObjectName("warningLabel")
-        warning.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        warning.setWordWrap(True)
-        root.addWidget(warning)
-        root.addSpacing(20)
-
-        root.addWidget(hr())
-        root.addSpacing(16)
-
-        
-        step1_lbl = QLabel(
-            "1.  Tap the  Receive  button on the Android app first."
-        )
-        step1_lbl.setObjectName("portalHint")
-        step1_lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        step1_lbl.setWordWrap(True)
-        root.addWidget(step1_lbl)
-        root.addSpacing(10)
-
-        step2_lbl = QLabel(
-            '2.  Then click Start Streaming below. When the screen-sharing '
-            'popup appears, select "TabletDisplay" and click Share.'
-        )
-        step2_lbl.setObjectName("portalHint")
-        step2_lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        step2_lbl.setWordWrap(True)
-        root.addWidget(step2_lbl)
-        root.addSpacing(20)
-
-        root.addStretch()
-
-        
-        back = QPushButton("← Back")
-        back.setObjectName("backBtn")
-        back.clicked.connect(on_back)
-
-        self._start_btn = QPushButton("▶  Start Streaming")
-        self._start_btn.setObjectName("primaryBtn")
-        self._start_btn.clicked.connect(self._validate_and_start)
-
-        row = QHBoxLayout()
-        row.setSpacing(20)
-        row.addStretch()
-        row.addWidget(back)
-        row.addWidget(self._start_btn)
-        row.addStretch()
-        root.addLayout(row)
-        root.addSpacing(20)
-
-    
-
-    def _on_res_changed(self, text: str):
-        self._custom_res_widget.setVisible(text == "Custom…")
-
-    def _on_fps_changed(self, text: str):
-        self._custom_fps_widget.setVisible(text == "Custom…")
-
-    
-
-    def _validate_and_start(self):
-        
-        if self._res_combo.currentText() == "Custom…":
-            try:
-                w = int(self._custom_w.text())
-                h = int(self._custom_h.text())
-            except ValueError:
-                QMessageBox.warning(
-                    self, "Invalid Resolution",
-                    "Please enter numeric values for width and height."
-                )
-                return
-            if not (500 <= w <= 4000):
-                QMessageBox.warning(
-                    self, "Invalid Resolution",
-                    f"Width must be between 500 and 4000.\nYou entered: {w}"
-                )
-                return
-            if not (500 <= h <= 4000):
-                QMessageBox.warning(
-                    self, "Invalid Resolution",
-                    f"Height must be between 500 and 4000.\nYou entered: {h}"
-                )
-                return
-
-        
-        if self._fps_combo.currentText() == "Custom…":
-            try:
-                fps = int(self._custom_fps_edit.text())
-            except ValueError:
-                QMessageBox.warning(
-                    self, "Invalid FPS",
-                    "Please enter a numeric value for FPS."
-                )
-                return
-            if not (24 <= fps <= 240):
-                QMessageBox.warning(
-                    self, "Invalid FPS",
-                    f"FPS must be between 24 and 240.\nYou entered: {fps}"
-                )
-                return
-
-        
-        try:
-            bitrate = int(self._bitrate_edit.text())
-        except ValueError:
-            QMessageBox.warning(self, "Invalid Bitrate", "Please enter a numeric value for Bitrate.")
-            return
-        if not (1000 <= bitrate <= 100000):
-            QMessageBox.warning(self, "Invalid Bitrate", "Bitrate must be between 1000 and 100000 kbps.")
-            return
-
-        self._on_start_cb()
-
-    
-
-    def selected_resolution(self) -> tuple[int, int]:
-        """Return (width, height) — handles Custom… selection."""
-        if self._res_combo.currentText() == "Custom…":
-            return int(self._custom_w.text()), int(self._custom_h.text())
-        text = self._res_combo.currentText()   
-        w, h = text.split("x")
-        return int(w), int(h)
-
-    def selected_fps(self) -> int:
-        if self._fps_combo.currentText() == "Custom…":
-            return int(self._custom_fps_edit.text())
-        return int(self._fps_combo.currentText())
-
-    def selected_bitrate(self) -> int:
-        return int(self._bitrate_edit.text())
-
-    def gnome_scale(self) -> str:
-        if hasattr(self, "_gnome_scale_combo"):
-            return self._gnome_scale_combo.currentText()
-        return "1.0"
-
-    def gnome_type(self) -> str:
-        if hasattr(self, "_display_type_combo"):
-            return self._display_type_combo.currentText()
-        return "Extend Right"
-
-
-class StreamingPage(QWidget):
-    """
-    Active streaming view.
-    Shows a live log box fed from both QProcess stdout streams.
-    """
-
-    def __init__(self, on_stop, on_configure, parent=None):
-        super().__init__(parent)
-        root = QVBoxLayout(self)
-        root.setContentsMargins(40, 36, 40, 36)
-        root.setSpacing(0)
-
-        
-        self._status_lbl = QLabel("Starting virtual monitor…")
-        self._status_lbl.setObjectName("streaming")
-        self._status_lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        root.addWidget(self._status_lbl)
-        root.addSpacing(6)
-
-        hint = QLabel(
-            "Live output from both processes is shown below."
-        )
-        hint.setObjectName("statusLbl")
-        hint.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        hint.setWordWrap(True)
-        root.addWidget(hint)
-        root.addSpacing(16)
-
-        
-        self.log = QPlainTextEdit()
-        self.log.setObjectName("logBox")
-        self.log.setReadOnly(True)
-        self.log.setMaximumBlockCount(800)
-        root.addWidget(self.log, stretch=1)
-        root.addSpacing(20)
-
-        
-        self._stop_btn = QPushButton("⏹  Stop Streaming")
-        self._stop_btn.setObjectName("stopBtn")
-        self._stop_btn.setEnabled(False)
-        self._stop_btn.clicked.connect(on_stop)
-
-        
-        self._configure_btn = QPushButton("⚙  Configure Display")
-        self._configure_btn.clicked.connect(on_configure)
-        self._configure_btn.setVisible(False)
-
-        row = QHBoxLayout()
-        row.addStretch()
-        row.addWidget(self._stop_btn)
-        row.addWidget(self._configure_btn)
-        row.addStretch()
-        root.addLayout(row)
-        root.addSpacing(16)
-
-    
-
-    def set_status(self, text: str):
-        self._status_lbl.setText(text)
-
-    def set_stop_enabled(self, enabled: bool):
-        self._stop_btn.setEnabled(enabled)
-
-    def set_configure_visible(self, visible: bool):
-        self._configure_btn.setVisible(visible)
-
-    def append_log(self, prefix: str, text: str):
-        """Append a labelled line to the log and auto-scroll."""
-        for line in text.splitlines():
-            if line.strip():
-                self.log.appendPlainText(f"[{prefix}] {line}")
-        self.log.moveCursor(QTextCursor.MoveOperation.End)
-
-    def clear_log(self):
-        self.log.clear()
-
-
-
-
-
-PAGE_MAIN      = 0
-PAGE_WIFI      = 1
-PAGE_USB1      = 2
-PAGE_USB2      = 3
-PAGE_STREAMING = 4
-
-
-def make_scrollable(widget: QWidget) -> QScrollArea:
-    scroll = QScrollArea()
-    scroll.setWidgetResizable(True)
-    scroll.setWidget(widget)
-    scroll.setFrameShape(QFrame.Shape.NoFrame)
-    scroll.setHorizontalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-    scroll.setVerticalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAsNeeded)
-    return scroll
-
-
-
-
-
-class MonitorizeWindow(QMainWindow):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("Monitorize")
-        self.setMinimumSize(760, 520)
-        self.resize(860, 580)
-
-        
-        import subprocess
-        subprocess.run(["killall", "-9", "gst-launch-1.0"], capture_output=True)
-        subprocess.run(["pkill", "-9", "-f", "Streamer_.*\\.py"], capture_output=True)
-
-        detected = detect_desktop_environment()
-        if detected:
-            self.detected_de = detected
-        else:
-            self.detected_de = self._ask_desktop_environment()
-        
-
-        self.initial_headless_monitors = []
-        if self.detected_de == "hyprland":
-            self.initial_headless_monitors = self._get_current_headless_monitors()
-            print(f"[Hyprland] Initial virtual monitors detected: {self.initial_headless_monitors}")
-        self.created_headless_monitor = None
-
-        
-        self.process_krfb:          QProcess | None = None
-        self.process_streamer:      QProcess | None = None
-        self.process_input_bridge:  QProcess | None = None
-
-        
-        self._proc_adb_dev:  QProcess | None = None
-        self._proc_adb_fwd:  QProcess | None = None
-        self._proc_adb_fwd2:  QProcess | None = None   
-
-        
-        self._countdown: int = 0
-        self._countdown_timer = QTimer(self)
-        self._countdown_timer.setInterval(1000)   
-        self._countdown_timer.timeout.connect(self._countdown_tick)
-
-        
-        self._stack = QStackedWidget()
-        self.setCentralWidget(self._stack)
-
-        self._page_main      = MainMenuPage(self._go_usb1, self._go_wifi)
-        self._page_wifi      = WifiPage(self._go_main, self._on_start_streaming_wifi, detected_de=self.detected_de)
-        self._page_usb1      = UsbStep1Page(self._go_main, self._on_connected)
-        self._page_usb2      = UsbStep2Page(self._go_usb1, self._on_start_streaming, detected_de=self.detected_de)
-        self._page_streaming = StreamingPage(self._on_stop_streaming, self._on_configure_display)
-
-        self._stack.addWidget(self._page_main)       
-        self._stack.addWidget(make_scrollable(self._page_wifi))  
-        self._stack.addWidget(self._page_usb1)       
-        self._stack.addWidget(make_scrollable(self._page_usb2))  
-        self._stack.addWidget(self._page_streaming)  
-
-        
-        self._page_main.update_de_badge(self.detected_de)
-
-        
-        
-        
-        try:
-            from zeroconf import ServiceInfo, Zeroconf
-            import socket
-            hostname = socket.gethostname()
-            self._zc = Zeroconf()
-            desc = {'name': hostname, 'port': 7110}
-            
-            
-            ip_addr = get_local_ip()
-            
-            self._info = ServiceInfo(
-                "_monitorize._tcp.local.",
-                f"{hostname}._monitorize._tcp.local.",
-                addresses=[socket.inet_aton(ip_addr)],
-                port=7110,
-                properties=desc,
-                server=f"{hostname}.local.",
-            )
-            self._zc.register_service(self._info)
-        except Exception as e:
-            print("Zeroconf registration failed:", e)
-
-        
-        
-        
-        self._tray = QSystemTrayIcon(self)
-        self._tray.setIcon(_make_tray_icon())
-        self._tray.setToolTip("Monitorize")
-
-        tray_menu = QMenu()
-        show_action = tray_menu.addAction("Show")
-        show_action.triggered.connect(self._restore_from_tray)
-        tray_menu.addSeparator()
-        quit_action = tray_menu.addAction("Quit")
-        quit_action.triggered.connect(self._quit_app)
-
-        self._tray.setContextMenu(tray_menu)
-        self._tray.activated.connect(self._tray_activated)
-        
-        self._tray.hide()
-
-    
-    
-    
-
-    def _ask_desktop_environment(self) -> str:
-        """
-        Show a QDialog asking the user to pick their DE.
-        Returns one of: "kde", "gnome", "hyprland", "sway".
-        Shows a WIP message and exits if the user picks "Other".
-        """
-        dlg = QDialog()
-        dlg.setWindowTitle("Select Desktop Environment")
-        dlg.setModal(True)
-
-        layout = QVBoxLayout(dlg)
-        layout.setSpacing(16)
-        layout.setContentsMargins(32, 28, 32, 28)
-
-        lbl = QLabel(
-            "Could not automatically detect your desktop environment.\n"
-            "Please select which one you are running:"
-        )
-        lbl.setWordWrap(True)
-        lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        layout.addWidget(lbl)
-
-        
-        row1 = QHBoxLayout()
-        row1.setSpacing(14)
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        kde_svg = os.path.join(script_dir, "assets", "svg", "kde-logo.svg")
-        gnome_svg = os.path.join(script_dir, "assets", "svg", "gnome-logo.svg")
-        hypr_svg = os.path.join(script_dir, "assets", "svg", "hyprland-logo.svg")
-
-        kde_btn   = QPushButton("  KDE")
-        if os.path.exists(kde_svg):
-            kde_btn.setIcon(QIcon(kde_svg))
-            kde_btn.setIconSize(QSize(20, 20))
-
-        gnome_btn = QPushButton("  GNOME")
-        if os.path.exists(gnome_svg):
-            gnome_btn.setIcon(QIcon(gnome_svg))
-            gnome_btn.setIconSize(QSize(20, 20))
-
-        row1.addWidget(kde_btn)
-        row1.addWidget(gnome_btn)
-        layout.addLayout(row1)
-
-        
-        row2 = QHBoxLayout()
-        row2.setSpacing(14)
-        hypr_btn  = QPushButton("  Hyprland")
-        if os.path.exists(hypr_svg):
-            hypr_btn.setIcon(QIcon(hypr_svg))
-            hypr_btn.setIconSize(QSize(20, 20))
-
-        sway_btn  = QPushButton("Sway")
-        row2.addWidget(hypr_btn)
-        row2.addWidget(sway_btn)
-        layout.addLayout(row2)
-
-        
-        other_btn = QPushButton("Other / Unsupported")
-        other_btn.setObjectName("backBtn")
-        layout.addWidget(other_btn)
-
-        dlg.setMinimumWidth(420)
-
-        chosen = [""]
-
-        def pick(value):
-            chosen[0] = value
-            dlg.accept()
-
-        kde_btn.clicked.connect(lambda: pick("kde"))
-        gnome_btn.clicked.connect(lambda: pick("gnome"))
-        hypr_btn.clicked.connect(lambda: pick("hyprland"))
-        sway_btn.clicked.connect(lambda: pick("sway"))
-        other_btn.clicked.connect(lambda: pick("other"))
-
-        dlg.exec()
-
-        if chosen[0] in ("other", ""):
-            msg = QMessageBox()
-            msg.setWindowTitle("Unsupported Desktop Environment")
-            msg.setText(
-                "Only KDE, GNOME, Hyprland, and Sway are supported.\n"
-                "Support for other desktop environments is a work in progress."
-            )
-            msg.setIcon(QMessageBox.Icon.Information)
-            msg.exec()
-            sys.exit(0)
-
-        return chosen[0]   
-
-    
-    
-    
-
-    def _go_main(self):
-        self._stack.setCurrentIndex(PAGE_MAIN)
-
-    def _go_wifi(self):
-        self._is_wifi = True
-        self._stack.setCurrentIndex(PAGE_WIFI)
-
-    def _go_usb1(self):
-        self._page_usb1.set_status("")
-        self._page_usb1.set_busy(False)
-        self._stack.setCurrentIndex(PAGE_USB1)
-
-    
-    
-    
-
-    def _on_connected(self):
-        self._page_usb1.set_busy(True)
-        self._page_usb1.set_status("Running adb devices…")
-
-        self._proc_adb_dev = QProcess(self)
-        self._proc_adb_dev.finished.connect(self._adb_devices_done)
-        self._proc_adb_dev.start("adb", ["devices"])
-
-    def _adb_devices_done(self, exit_code, _status):
-        if exit_code != 0:
-            self._page_usb1.set_status("Error: adb devices failed. Is ADB installed?")
-            self._page_usb1.set_busy(False)
-            return
-
-        self._page_usb1.set_status("Forwarding port tcp:7110…")
-        self._proc_adb_fwd = QProcess(self)
-        self._proc_adb_fwd.finished.connect(self._adb_forward_done)
-        self._proc_adb_fwd.start("adb", ["forward", "tcp:7110", "tcp:7110"])
-
-    def _adb_forward_done(self, exit_code, _status):
-        if exit_code != 0:
-            self._page_usb1.set_status("Error: Port forward failed. Is a device connected?")
-            self._page_usb1.set_busy(False)
-            return
-
-        
-        
-        self._page_usb1.set_status("Setting up reverse proxy tcp:7111 (touch)…")
-        self._proc_adb_fwd2 = QProcess(self)
-        self._proc_adb_fwd2.finished.connect(self._adb_forward2_done)
-        self._proc_adb_fwd2.start("adb", ["reverse", "tcp:7111", "tcp:7111"])
-
-    def _adb_forward2_done(self, exit_code, _status):
-        
-        if exit_code != 0:
-            self._page_usb1.set_status("Warning: tcp:7111 reverse failed — touch disabled")
-        else:
-            self._page_usb1.set_status("Device ready!")
-        self._page_usb1.set_busy(False)
-        QTimer.singleShot(600, lambda: self._stack.setCurrentIndex(PAGE_USB2))
-
-    
-    
-    
-
-    def _on_start_streaming_wifi(self):
-        self._is_wifi = True
-        self._do_start_streaming(self._page_wifi)
-
-    def _on_start_streaming(self):
-        self._is_wifi = False
-        self._do_start_streaming(self._page_usb2)
-
-    def _do_start_streaming(self, config_page):
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        self._page_streaming.clear_log()
-
-        env = QProcessEnvironment.systemEnvironment()
-        env.insert("PYTHONUNBUFFERED", "1")
-        self._script_dir = script_dir
-        self._env        = env
-
-        self._page_streaming.set_stop_enabled(False)
-        self._page_streaming.set_configure_visible(self.detected_de == "hyprland")
-        self._stack.setCurrentIndex(PAGE_STREAMING)
-
-        width, height = config_page.selected_resolution()
-        fps = config_page.selected_fps()
-        bitrate = config_page.selected_bitrate()
-        self._stream_width  = width
-        self._stream_height = height
-        self._stream_fps    = fps
-        self._stream_bitrate = bitrate
-        import subprocess
-        subprocess.run(["killall", "-9", "gst-launch-1.0"], capture_output=True)
-        subprocess.run(["pkill", "-9", "-f", "Streamer_.*\\.py"], capture_output=True)
-        
-        if self._is_wifi:
-            
-            subprocess.run(["adb", "forward", "--remove", "tcp:7110"], capture_output=True)
-            subprocess.run(["adb", "reverse", "--remove", "tcp:7111"], capture_output=True)
-
-        display_type = config_page.gnome_type()  
-
-        if self.detected_de == "kde":
-            if display_type == "Mirror":
-                self._page_streaming.set_status("Launching streamer (Mirror mode)…")
-                self._launch_streamer()
-            else:
-                self._page_streaming.set_status("Starting virtual monitor…  5")
-                self.process_krfb = QProcess(self)
-                self.process_krfb.setWorkingDirectory(script_dir)
-                self.process_krfb.setProcessEnvironment(env)
-                self.process_krfb.setProcessChannelMode(QProcess.ProcessChannelMode.MergedChannels)
-                self.process_krfb.readyReadStandardOutput.connect(self._read_krfb)
-                self.process_krfb.finished.connect(
-                    lambda code, _: self._page_streaming.append_log("KRFB", f"Process exited (code {code})")
-                )
-                import subprocess
-                subprocess.run(["killall", "krfb-virtualmonitor"], capture_output=True)
-
-                self.process_krfb.start("krfb-virtualmonitor", [
-                    "--resolution", f"{width}x{height}",
-                    "--name",       "TabletDisplay",
-                    "--password",   "test123",
-                    "--port",       "5900",
-                ])
-                self._countdown = 1
-                self._countdown_timer.start()
-        elif self.detected_de == "hyprland":
-            if display_type == "Mirror":
-                self._page_streaming.set_status("Launching streamer (Mirror mode)…")
-                self._launch_streamer()
-            else:
-                self._page_streaming.set_status("Setting up virtual monitor on Hyprland…")
-                import subprocess
-                old_monitors = set(self._get_current_headless_monitors())
-                subprocess.run(["hyprctl", "output", "create", "headless"], capture_output=True)
-                
-                new_monitors = set(self._get_current_headless_monitors())
-                diff = new_monitors - old_monitors
-                if diff:
-                    new_name = list(diff)[0]
-                else:
-                    new_name = "HEADLESS-1"
-                
-                self.created_headless_monitor = new_name
-                
-                subprocess.run(["hyprctl", "keyword", "monitor", f"{new_name},{width}x{height}@{fps},auto,1"], capture_output=True)
-                subprocess.run(["hyprctl", "eval", f"hl.monitor({{ output = '{new_name}', mode = '{width}x{height}@{fps}', position = 'auto', scale = 1.0 }})"], capture_output=True)
-                print(f"[Hyprland] Created new headless monitor: {new_name} at {width}x{height}@{fps}")
-                self._page_streaming.append_log("STREAMER", f"Created new headless monitor: {new_name} at {width}x{height}@{fps}")
-                self._launch_streamer()
-        else:
-            self._page_streaming.set_status("Launching streamer…")
-            self._launch_streamer()
-
-    def _countdown_tick(self):
-        """Called every 1 s by _countdown_timer. Starts the streamer at 0."""
-        self._countdown -= 1
-
-        if self._countdown > 0:
-            self._page_streaming.set_status(
-                f"Starting virtual monitor…  {self._countdown}"
-            )
-            return
-        
-        self._countdown_timer.stop()
-        self._launch_streamer()
-
-    def _launch_streamer(self):
-        """Spawn the correct DE-specific streamer script as a QProcess."""
-        self.process_streamer = QProcess(self)
-        self.process_streamer.setWorkingDirectory(self._script_dir)
-        self.process_streamer.setProcessEnvironment(self._env)
-        self.process_streamer.setProcessChannelMode(
-            QProcess.ProcessChannelMode.MergedChannels
-        )
-        self.process_streamer.readyReadStandardOutput.connect(self._read_streamer)
-        self.process_streamer.finished.connect(self._on_streamer_finished)
-        
-        _streamer_map = {
-            "kde":      "Streamer_kde_usb.py",
-            "gnome":    "Streamer_gnome_wifi.py" if self._is_wifi else "Streamer_gnome_usb.py",
-            "hyprland": "Streamer_hyprland_wifi.py" if self._is_wifi else "Streamer_hyprland_usb.py",
-            "sway":     "Streamer_sway_usb.py",
-        }
-        streamer_script = _streamer_map.get(self.detected_de, "Streamer_kde_usb.py")
-
-        args = [
-            streamer_script,
-            str(self._stream_width),
-            str(self._stream_height),
-            str(self._stream_fps),
-            str(self._stream_bitrate),
-        ]
-        if self._is_wifi:
-            args.append("wifi")
-        else:
-            args.append("usb")
-
-        
-        if self.detected_de == "hyprland":
-            if getattr(self, "created_headless_monitor", None):
-                args.append(self.created_headless_monitor)
-            else:
-                args.append("mirror")
-
-        if self.detected_de == "gnome":
-            active_page = self._page_wifi if self._is_wifi else self._page_usb2
-            args.append(active_page.gnome_scale())
-            args.append(active_page.gnome_type().replace(" ", "_"))
-
-        self.process_streamer.start("python3", args)
-
-        
-        
-        
-        
-        if self.detected_de in ("kde", "gnome"):
-            QTimer.singleShot(400, self._launch_input_bridge)
-        elif self.detected_de == "hyprland":
-            self._input_bridge_launched = False
-            self._streamer_buffer = ""
-
-        self._page_streaming.set_status("Status: Streaming…")
-        self._page_streaming.set_stop_enabled(True)
-
-    def _launch_input_bridge(self):
-        """Spawn input_bridge.py — listens on port 7111 for Android touch/pen events."""
-        self.process_input_bridge = QProcess(self)
-        self.process_input_bridge.setWorkingDirectory(self._script_dir)
-        self.process_input_bridge.setProcessEnvironment(self._env)
-        self.process_input_bridge.setProcessChannelMode(
-            QProcess.ProcessChannelMode.MergedChannels
-        )
-        self.process_input_bridge.readyReadStandardOutput.connect(self._read_input_bridge)
-        self.process_input_bridge.finished.connect(self._on_input_bridge_finished)
-        args = [
-            os.path.join(self._script_dir, "touch_daemon.py"),
-            str(self._stream_width),
-            str(self._stream_height),
-        ]
-        if getattr(self, "_is_wifi", False):
-            args.append("--wifi")
-        self.process_input_bridge.start("python3", args)
-
-        if self.detected_de == "hyprland":
-            self._page_streaming.set_status(
-                "Touch service starting via uinput…"
-            )
-        else:
-            self._page_streaming.set_status(
-                "Touch service starting… Watch for 'Allow Remote Control' popup and click Allow"
-            )
-
-    
-    
-    
-
-    def _on_streamer_finished(self, code, _status):
-        """Handle streamer process exit. On GNOME, auto-restart after crashes."""
-        self._page_streaming.append_log("STREAMER", f"Process exited (code {code})")
-
-        
-        
-        
-        
-        if (self.detected_de == "gnome"
-                and code != 0
-                and self._stack.currentIndex() == PAGE_STREAMING):
-            self._page_streaming.append_log(
-                "STREAMER",
-                "⟳  GNOME streamer crashed — auto-restarting in 2s…"
-            )
-            self._page_streaming.set_status(
-                "⟳  Stream reconnecting after display config change…"
-            )
-            QTimer.singleShot(2000, self._gnome_restart_streamer)
-
-    def _on_input_bridge_finished(self, code, _status):
-        """Handle touch daemon process exit."""
-        self._page_streaming.append_log("INPUT", f"Bridge exited (code {code})")
-
-        if code == 0 and self._stack.currentIndex() == PAGE_STREAMING:
-            
-            
-            
-            self._page_streaming.append_log(
-                "INPUT",
-                "ℹ️  Touch input not available — streaming continues without touch."
-            )
-
-    def _gnome_restart_streamer(self):
-        """Auto-restart the GNOME streamer after a crash."""
-        if self._stack.currentIndex() != PAGE_STREAMING:
-            return   
-        
-        import subprocess
-        subprocess.run(["killall", "-9", "gst-launch-1.0"], capture_output=True)
-        self._launch_streamer()
-        self._page_streaming.set_status("Status: Streaming…  (restarted)")
-
-    def _gnome_restart_input_bridge(self):
-        """Auto-restart the touch daemon after a crash on GNOME."""
-        if self._stack.currentIndex() != PAGE_STREAMING:
-            return   
-        self._launch_input_bridge()
-
-    
-    
-    
-
-    def _read_krfb(self):
-        if self.process_krfb is None:
-            return
-        raw = bytes(self.process_krfb.readAllStandardOutput()).decode(
-            "utf-8", errors="replace"
-        )
-        self._page_streaming.append_log("KRFB", raw)
-
-    def _read_streamer(self):
-        if self.process_streamer is None:
-            return
-        raw = bytes(self.process_streamer.readAllStandardOutput()).decode(
-            "utf-8", errors="replace"
-        )
-        self._page_streaming.append_log("STREAMER", raw)
-
-        
-        
-        
-        
-        if self.detected_de == "hyprland":
-            if not getattr(self, "_input_bridge_launched", False):
-                if not hasattr(self, "_streamer_buffer"):
-                    self._streamer_buffer = ""
-                self._streamer_buffer += raw
-                if "[Portal] Got PipeWire node=" in self._streamer_buffer:
-                    self._input_bridge_launched = True
-                    
-                    
-                    QTimer.singleShot(500, self._launch_input_bridge)
-
-    def _read_input_bridge(self):
-        if self.process_input_bridge is None:
-            return
-        raw = bytes(self.process_input_bridge.readAllStandardOutput()).decode(
-            "utf-8", errors="replace"
-        )
-        self._page_streaming.append_log("INPUT", raw)
-
-    
-    
-    
-
-    def _on_stop_streaming(self):
-        self._kill_stream_procs()
-        self._go_main()
-
-    def _on_configure_display(self):
-        """Run nwg-displays for Hyprland display configuration."""
-        import shutil
-        cmd = "nwg-displays"
-        if shutil.which(cmd) is None:
-            QMessageBox.warning(
-                self,
-                "nwg-displays Not Installed",
-                "nwg-displays is not installed."
-            )
-            return
-
-        try:
-            QProcess.startDetached(cmd)
-        except Exception as e:
-            if "not found" in str(e).lower() or isinstance(e, FileNotFoundError):
-                QMessageBox.warning(
-                    self,
-                    "nwg-displays Not Installed",
-                    "nwg-displays is not installed."
-                )
-            else:
-                QMessageBox.warning(
-                    self,
-                    "Error",
-                    f"Failed to run nwg-displays:\n{str(e)}"
-                )
-
-    def _kill_stream_procs(self):
-        
-        self._countdown_timer.stop()
-        self._countdown = 0
-
-        for proc in (self.process_krfb, self.process_streamer, self.process_input_bridge):
-            if proc is not None and proc.state() != QProcess.ProcessState.NotRunning:
-                proc.terminate()
-                if not proc.waitForFinished(3000):
-                    proc.kill()
-        self.process_krfb          = None
-        self.process_streamer      = None
-        self.process_input_bridge  = None
-
-        
-        if self.detected_de == "hyprland" and getattr(self, "created_headless_monitor", None):
-            print(f"[Hyprland] Removing created headless monitor: {self.created_headless_monitor}")
-            import subprocess
-            subprocess.run(["hyprctl", "output", "remove", self.created_headless_monitor], capture_output=True)
-            self.created_headless_monitor = None
-
-    def _get_current_headless_monitors(self):
-        """Query and return a list of currently existing headless monitor names."""
-        import subprocess, json, re
-        headless_names = []
-        
-        try:
-            res = subprocess.run(["hyprctl", "monitors", "all", "-j"], capture_output=True, text=True)
-            if res.returncode == 0:
-                monitors = json.loads(res.stdout)
-                for mon in monitors:
-                    name = mon.get("name", "")
-                    if name.startswith("HEADLESS"):
-                        headless_names.append(name)
-                return headless_names
-        except Exception:
-            pass
-
-        
-        try:
-            res = subprocess.run(["hyprctl", "monitors", "all"], capture_output=True, text=True)
-            if res.returncode == 0:
-                matches = re.findall(r"\bHEADLESS-\d+\b", res.stdout)
-                headless_names = list(set(matches))
-        except Exception:
-            pass
-        return headless_names
-
-    
-    
-    
-
-    def _tray_activated(self, reason):
-        """Restore the window when the tray icon is double-clicked."""
-        if reason == QSystemTrayIcon.ActivationReason.DoubleClick:
-            self._restore_from_tray()
-
-    def _restore_from_tray(self):
-        self._tray.hide()
-        self.showNormal()
-        self.activateWindow()
-
-    def _quit_app(self):
-        """Hard quit from the tray menu — always terminates processes."""
-        self._kill_stream_procs()
-        if hasattr(self, '_zc') and hasattr(self, '_info'):
-            try:
-                self._zc.unregister_service(self._info)
-                self._zc.close()
-            except Exception: pass
-        QApplication.quit()
-
-    
-    
-    
-
-    def closeEvent(self, event):
-        if self._page_main.tray_checkbox.isChecked():
-            
-            event.ignore()
-            self.hide()
-            self._tray.show()
-            self._tray.showMessage(
-                "Monitorize",
-                "Running in the background. Double-click the tray icon to restore.",
-                QSystemTrayIcon.MessageIcon.Information,
-                3000,
-            )
-        else:
-            self._kill_stream_procs()
-            if hasattr(self, '_zc') and hasattr(self, '_info'):
-                try:
-                    self._zc.unregister_service(self._info)
-                    self._zc.close()
-                except Exception: pass
-            event.accept()
-
-
-
-
-
-
-def main():
-    app = QApplication(sys.argv)
-    app.setApplicationName("Monitorize")
-    app.setStyleSheet(DARK_QSS)
-
-    palette = QPalette()
-    palette.setColor(QPalette.ColorRole.Window,          QColor("#0c0d14"))
-    palette.setColor(QPalette.ColorRole.WindowText,      QColor("#d4d6f0"))
-    palette.setColor(QPalette.ColorRole.Base,            QColor("#12142a"))
-    palette.setColor(QPalette.ColorRole.AlternateBase,   QColor("#16182a"))
-    palette.setColor(QPalette.ColorRole.Button,          QColor("#16182a"))
-    palette.setColor(QPalette.ColorRole.ButtonText,      QColor("#d4d6f0"))
-    palette.setColor(QPalette.ColorRole.Highlight,       QColor("#3538b0"))
-    palette.setColor(QPalette.ColorRole.HighlightedText, QColor("#ffffff"))
-    app.setPalette(palette)
-
-    win = MonitorizeWindow()
-    win.show()
-    sys.exit(app.exec())
-
+from gui.main_window import main
 
 if __name__ == "__main__":
     main()
